Make SubtileGridSizer.from_tile_params public API (#124)
This PR moves `SubtileGridSizer._from_tile_params` to `SubtileGridSizer.from_tile_params`.

Users were finding it unclear whether SubtileGridSizer is initialized using the shape of the subtile or of the full tile. Providing this class method should make it clearer that the main method accepts subtile shapes. It has also proven to be a useful way to create a SubtileGridSizer in private code, and should not need to change..